+++ src/vite-env.d.ts
@@ -6,6 +6,7 @@ declare global {
   const __APP_VERSION__: string;
 
   interface Window {
+    _qdnBase?: unknown;
     _qdnIdentifier?: unknown;
     _qdnName?: unknown;
     _qdnPath?: unknown;